shorten ipc completion events
